+++ packages/bot-api/CHANGELOG.md
@@ -3,6 +3,110 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.4](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@7.0.3...@wireapp/bot-api@7.0.4) (2019-11-22)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [7.0.3](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@7.0.2...@wireapp/bot-api@7.0.3) (2019-11-21)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [7.0.2](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@7.0.1...@wireapp/bot-api@7.0.2) (2019-11-21)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [7.0.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@7.0.0...@wireapp/bot-api@7.0.1) (2019-11-21)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+# [7.0.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@6.0.16...@wireapp/bot-api@7.0.0) (2019-11-21)
+
+
+### Code Refactoring
+
+* **api-client,bot-api,changelog-bot,cli-client,commons,core,store-engine-bro-fs,store-engine-dexie,store-engine-fs,store-engine-sqleet,store-engine-web-storage,travis-bot:** Remove store engine from API client ([#2558](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/issues/2558)) ([6b8ba89](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/commit/6b8ba892c85ca43cd498a7b3d56a20a31f8578a4))
+
+
+### BREAKING CHANGES
+
+* **api-client,bot-api,changelog-bot,cli-client,commons,core,store-engine-bro-fs,store-engine-dexie,store-engine-fs,store-engine-sqleet,store-engine-web-storage,travis-bot:** Store engine has been removed from "api-client" and needs to be initialized with a provider function within the "core" construction.
+
+
+
+
+
+## [6.0.16](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@6.0.15...@wireapp/bot-api@6.0.16) (2019-11-21)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [6.0.15](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@6.0.14...@wireapp/bot-api@6.0.15) (2019-11-20)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [6.0.14](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@6.0.13...@wireapp/bot-api@6.0.14) (2019-11-19)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [6.0.13](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@6.0.12...@wireapp/bot-api@6.0.13) (2019-11-12)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [6.0.12](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@6.0.11...@wireapp/bot-api@6.0.12) (2019-11-11)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [6.0.11](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@6.0.10...@wireapp/bot-api@6.0.11) (2019-11-11)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
+## [6.0.10](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@6.0.9...@wireapp/bot-api@6.0.10) (2019-11-11)
+
+**Note:** Version bump only for package @wireapp/bot-api
+
+
+
+
+
 ## [6.0.9](https://github.com/wireapp/wire-web-packages/tree/master/packages/bot-api/compare/@wireapp/bot-api@6.0.8...@wireapp/bot-api@6.0.9) (2019-11-06)
 
 **Note:** Version bump only for package @wireapp/bot-api