+++ packages/cryptobox/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.0.7](https://github.com/wireapp/wire-web-packages/tree/master/packages/cryptobox/compare/@wireapp/cryptobox@12.0.6...@wireapp/cryptobox@12.0.7) (2019-11-22)
+
+**Note:** Version bump only for package @wireapp/cryptobox
+
+
+
+
+
+## [12.0.6](https://github.com/wireapp/wire-web-packages/tree/master/packages/cryptobox/compare/@wireapp/cryptobox@12.0.5...@wireapp/cryptobox@12.0.6) (2019-11-21)
+
+**Note:** Version bump only for package @wireapp/cryptobox
+
+
+
+
+
+## [12.0.5](https://github.com/wireapp/wire-web-packages/tree/master/packages/cryptobox/compare/@wireapp/cryptobox@12.0.4...@wireapp/cryptobox@12.0.5) (2019-11-21)
+
+**Note:** Version bump only for package @wireapp/cryptobox
+
+
+
+
+
+## [12.0.4](https://github.com/wireapp/wire-web-packages/tree/master/packages/cryptobox/compare/@wireapp/cryptobox@12.0.3...@wireapp/cryptobox@12.0.4) (2019-11-21)
+
+**Note:** Version bump only for package @wireapp/cryptobox
+
+
+
+
+
+## [12.0.3](https://github.com/wireapp/wire-web-packages/tree/master/packages/cryptobox/compare/@wireapp/cryptobox@12.0.2...@wireapp/cryptobox@12.0.3) (2019-11-11)
+
+**Note:** Version bump only for package @wireapp/cryptobox
+
+
+
+
+
+## [12.0.2](https://github.com/wireapp/wire-web-packages/tree/master/packages/cryptobox/compare/@wireapp/cryptobox@12.0.1...@wireapp/cryptobox@12.0.2) (2019-11-11)
+
+**Note:** Version bump only for package @wireapp/cryptobox
+
+
+
+
+
 ## [12.0.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/cryptobox/compare/@wireapp/cryptobox@12.0.0...@wireapp/cryptobox@12.0.1) (2019-11-06)
 
 **Note:** Version bump only for package @wireapp/cryptobox